+++ src/by/it/chetovich/JD03_01/DB_creation_java/CN.java
@@ -5,7 +5,7 @@ package by.it.chetovich.JD03_01.DB_creation_java;
  */
 public class CN {
 
-    public static final String URL_DB = "jdbc:mysql://127.0.0.1:3306/searchhost_java";
+    public static final String URL_DB = "jdbc:mysql://127.0.0.1:2016/searchhost_java";
     public static final String USER_DB = "root";
-    public static final String PASSWORD_DB = "yultosik";
+    public static final String PASSWORD_DB = "";
 }
